// apps/admin/src/features/business-settings/model/types.ts
export type BusinessFormValues = {
  name: string;
  description: string;
  address: string;
  phone: string;
  email: string;
  website: string;
  telegram: string;
  vk: string;
  youtube: string;
};

export const defaultBusinessFormValues: BusinessFormValues = {
  name: '',
  description: '',
  address: '',
  phone: '',
  email: '',
  website: '',
  telegram: '',
  vk: '',
  youtube: '',
};